+++ test/src/test_species_species.cpp
@@ -3,6 +3,7 @@
 #include <species/Species.hpp>
 #include <species/Species.cpp>
 #include <species/AttributeFactory.cpp>
+#include <types/Propability.cpp>
 
 TEST_CASE("fixed")
 {
@@ -12,12 +13,12 @@ TEST_CASE("fixed")
 
     auto const a = s.create("S");
     REQUIRE(a.state == "S");
-    REQUIRE(a.new_contact_rate == 1);
-    REQUIRE(a.loose_contact_rate == 2);
+    REQUIRE(static_cast<double> (a.new_contact_rate) == 1);
+    REQUIRE(static_cast<double> (a.loose_contact_rate) == 2);
 
     auto const b = s.create("I");
     REQUIRE(b.state == "I");
-    REQUIRE(b.new_contact_rate == 3);
-    REQUIRE(b.loose_contact_rate == 4);
+    REQUIRE(static_cast<double> (b.new_contact_rate) == 3);
+    REQUIRE(static_cast<double> (b.loose_contact_rate) == 4);
 }
